Add hash_set perf test
