+++ tests/perf/hashset/expected
@@ -1,0 +1,1 @@
+4 successfully verified harnesses, 0 failures, 4 total
